web_app/requirements.txt
Flask>=3.1,<4
Werkzeug>=3.1.0
Pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.0
torch==2.3.0
torchvision==0.18.0
requests==2.32.3
flask-socketio==5.3.6
tqdm==4.66.0
ultralytics>=8.0.0
imageio-ffmpeg
